ui changes
- high contract mode
- files seperation
- general fixes
