+++ src/electron/windows/styles/settings.css
@@ -1,0 +1,209 @@
+/* Settings Window Styles */
+* { 
+  margin: 0; 
+  padding: 0; 
+  box-sizing: border-box; 
+}
+
+body {
+  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', sans-serif;
+  padding: 2rem;
+  background: #f5f5f5;
+  color: #333;
+}
+
+[data-theme="dark"] body {
+  background: #1a1a1a;
+  color: #e0e0e0;
+}
+
+[data-theme="high-contrast"] body {
+  background: #000000;
+  color: #ffff00;
+}
+
+.container {
+  max-width: 500px;
+  margin: 0 auto;
+}
+
+h1 {
+  margin-bottom: 2rem;
+  font-size: 1.75rem;
+}
+
+.setting-item {
+  background: white;
+  padding: 1.5rem;
+  border-radius: 12px;
+  margin-bottom: 1rem;
+  box-shadow: 0 2px 8px rgba(0,0,0,0.1);
+}
+
+[data-theme="dark"] .setting-item {
+  background: #2a2a2a;
+  box-shadow: 0 2px 8px rgba(0,0,0,0.3);
+}
+
+[data-theme="high-contrast"] .setting-item {
+  background: #000000;
+  border: 2px solid #ffff00;
+  box-shadow: 0 0 0 1px #ffff00, 0 2px 8px rgba(255,255,0,0.5);
+}
+
+.setting-label {
+  font-weight: 600;
+  margin-bottom: 1rem;
+  display: block;
+}
+
+.theme-select {
+  width: 100%;
+}
+
+.theme-select select {
+  width: 100%;
+  padding: 0.75rem 1rem;
+  border: 2px solid #667eea;
+  border-radius: 8px;
+  background: white;
+  color: #667eea;
+  font-size: 1rem;
+  cursor: pointer;
+  transition: all 0.2s;
+  appearance: none;
+  background-image: url("data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' width='12' height='12' viewBox='0 0 12 12'%3E%3Cpath fill='%23667eea' d='M6 9L1 4h10z'/%3E%3C/svg%3E");
+  background-repeat: no-repeat;
+  background-position: right 1rem center;
+  padding-right: 2.5rem;
+}
+
+[data-theme="dark"] .theme-select select {
+  background-color: #2a2a2a;
+  color: #8a9aff;
+  border-color: #8a9aff;
+  background-image: url("data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' width='12' height='12' viewBox='0 0 12 12'%3E%3Cpath fill='%238a9aff' d='M6 9L1 4h10z'/%3E%3C/svg%3E");
+}
+
+[data-theme="high-contrast"] .theme-select select {
+  background-color: #000000;
+  color: #ffff00;
+  border-color: #ffff00;
+  border-width: 2px;
+  box-shadow: 0 0 0 1px #ffff00;
+  background-image: url("data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' width='12' height='12' viewBox='0 0 12 12'%3E%3Cpath fill='%23ffff00' d='M6 9L1 4h10z'/%3E%3C/svg%3E");
+}
+
+.theme-select select:hover {
+  border-color: #5568d3;
+  box-shadow: 0 2px 8px rgba(102, 126, 234, 0.2);
+}
+
+[data-theme="dark"] .theme-select select:hover {
+  border-color: #9aafff;
+  box-shadow: 0 2px 8px rgba(138, 154, 255, 0.3);
+}
+
+[data-theme="high-contrast"] .theme-select select:hover {
+  border-color: #ffff00;
+  box-shadow: 0 0 0 1px #ffff00, 0 2px 8px rgba(255, 255, 0, 0.4);
+}
+
+.theme-select select:focus {
+  outline: none;
+  border-color: #667eea;
+  box-shadow: 0 0 0 3px rgba(102, 126, 234, 0.1);
+}
+
+[data-theme="dark"] .theme-select select:focus {
+  border-color: #8a9aff;
+  box-shadow: 0 0 0 3px rgba(138, 154, 255, 0.2);
+}
+
+[data-theme="high-contrast"] .theme-select select:focus {
+  border-color: #ffff00;
+  box-shadow: 0 0 0 1px #ffff00, 0 0 0 3px rgba(255, 255, 0, 0.3);
+}
+
+.slicer-path {
+  display: flex;
+  flex-direction: column;
+  gap: 0.75rem;
+}
+
+.slicer-path-display {
+  padding: 0.75rem;
+  background: #f8f9fa;
+  border: 2px solid #e9ecef;
+  border-radius: 8px;
+  font-size: 0.875rem;
+  color: #666;
+  word-break: break-all;
+  min-height: 2.5rem;
+  display: flex;
+  align-items: center;
+}
+
+[data-theme="dark"] .slicer-path-display {
+  background: #2a2a2a;
+  border-color: rgba(255, 255, 255, 0.2);
+  color: rgba(255, 255, 255, 0.7);
+}
+
+[data-theme="high-contrast"] .slicer-path-display {
+  background: #000000;
+  border-color: #ffff00;
+  border-width: 2px;
+  box-shadow: 0 0 0 1px #ffff00;
+  color: #ffff00;
+}
+
+.slicer-path-display.empty {
+  color: #999;
+  font-style: italic;
+}
+
+[data-theme="dark"] .slicer-path-display.empty {
+  color: rgba(255, 255, 255, 0.4);
+}
+
+[data-theme="high-contrast"] .slicer-path-display.empty {
+  color: #ffff00;
+  opacity: 0.7;
+}
+
+.slicer-path button {
+  padding: 0.75rem 1.5rem;
+  border: 2px solid #667eea;
+  border-radius: 8px;
+  background: #667eea;
+  color: white;
+  font-size: 0.95rem;
+  cursor: pointer;
+  transition: all 0.2s;
+  font-weight: 500;
+}
+
+[data-theme="dark"] .slicer-path button {
+  background: #8a9aff;
+  border-color: #8a9aff;
+  color: #1a1a1a;
+}
+
+[data-theme="high-contrast"] .slicer-path button {
+  background: #ffff00;
+  border-color: #ffff00;
+  border-width: 2px;
+  box-shadow: 0 0 0 1px #ffff00;
+  color: #000000;
+}
+
+.slicer-path button:hover {
+  transform: translateY(-2px);
+  box-shadow: 0 4px 12px rgba(102, 126, 234, 0.3);
+}
+
+.slicer-path button:active {
+  transform: translateY(0);
+}
+
