+++ src/ui/styles/index.css
@@ -1,0 +1,11 @@
+/* Main Styles Index - Imports all styles */
+@import './base.css';
+@import './components/header.css';
+@import './components/image-preview.css';
+@import './components/settings.css';
+@import './components/controls.css';
+@import './components/popup.css';
+@import './themes/light.css';
+@import './themes/dark.css';
+@import './themes/high-contrast.css';
+
